+++ modules/Core/src/main/java/org/terasology/logic/health/HealthAuthoritySystem.java
@@ -37,12 +37,9 @@ import org.terasology.entitySystem.systems.UpdateSubscriberSystem;
 import org.terasology.logic.characters.CharacterSoundComponent;
 import org.terasology.logic.characters.CharacterSoundSystem;
 import org.terasology.logic.characters.events.AttackEvent;
-import org.terasology.logic.characters.events.HorizontalCollisionEvent;
-import org.terasology.logic.characters.events.VerticalCollisionEvent;
 import org.terasology.logic.inventory.ItemComponent;
 import org.terasology.logic.players.event.OnPlayerRespawnedEvent;
 import org.terasology.math.TeraMath;
-import org.terasology.math.geom.Vector3f;
 import org.terasology.registry.In;
 import org.terasology.utilities.random.FastRandom;
 import org.terasology.utilities.random.Random;
@@ -225,54 +222,6 @@ public class HealthAuthoritySystem extends BaseComponentSystem implements Update
     @ReceiveEvent(components = {HealthComponent.class})
     public void onHeal(DoHealEvent event, EntityRef entity) {
         checkHeal(entity, event.getAmount(), event.getInstigator());
-    }
-
-    // TODO: This should be in a separate system
-    @ReceiveEvent(components = {HealthComponent.class})
-    public void onLand(VerticalCollisionEvent event, EntityRef entity) {
-        HealthComponent health = entity.getComponent(HealthComponent.class);
-        float speed = Math.abs(event.getVelocity().y);
-
-        if (speed > health.fallingDamageSpeedThreshold) {
-            int damage = (int) ((speed - health.fallingDamageSpeedThreshold) * health.excessSpeedDamageMultiplier);
-            if (damage > 0) {
-                checkDamage(entity, damage, EngineDamageTypes.PHYSICAL.get(), EntityRef.NULL, EntityRef.NULL);
-            }
-        }
-    }
-
-    @ReceiveEvent(components = {HealthComponent.class})
-    public void onCrash(HorizontalCollisionEvent event, EntityRef entity) {
-        HealthComponent health = entity.getComponent(HealthComponent.class);
-
-        Vector3f vel = new Vector3f(event.getVelocity());
-        vel.y = 0;
-        float speed = vel.length();
-
-        if (speed > health.horizontalDamageSpeedThreshold) {
-            int damage = (int) ((speed - health.horizontalDamageSpeedThreshold) * health.excessSpeedDamageMultiplier);
-            if (damage > 0) {
-                checkDamage(entity, damage, EngineDamageTypes.PHYSICAL.get(), EntityRef.NULL, EntityRef.NULL);
-            }
-        }
-    }
-
-    @ReceiveEvent
-    public void onCrash(HorizontalCollisionEvent event, EntityRef entity, CharacterSoundComponent characterSounds, HealthComponent healthComponent) {
-        Vector3f horizVelocity = new Vector3f(event.getVelocity());
-        horizVelocity.y = 0;
-        float velocity = horizVelocity.length();
-
-        if (velocity > healthComponent.horizontalDamageSpeedThreshold) {
-            if (characterSounds.lastSoundTime + CharacterSoundSystem.MIN_TIME < time.getGameTimeInMs()) {
-                StaticSound sound = random.nextItem(characterSounds.landingSounds);
-                if (sound != null) {
-                    entity.send(new PlaySoundEvent(sound, characterSounds.landingVolume));
-                    characterSounds.lastSoundTime = time.getGameTimeInMs();
-                    entity.saveComponent(characterSounds);
-                }
-            }
-        }
     }
 
     @ReceiveEvent
